Handlers exceptions

+++ src/main/java/com/djalves424/ProjetoClientCRUD/services/ClientService.java
@@ -3,10 +3,15 @@ package com.djalves424.ProjetoClientCRUD.services;
 import com.djalves424.ProjetoClientCRUD.dto.ClientDTO;
 import com.djalves424.ProjetoClientCRUD.entities.Client;
 import com.djalves424.ProjetoClientCRUD.entities.repositories.ClientRepository;
+import com.djalves424.ProjetoClientCRUD.services.exceptions.DatabaseException;
+import com.djalves424.ProjetoClientCRUD.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +22,7 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
-        Client client = repository.findById(id).get();
+        Client client = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Cliente inexistente"));
         return new ClientDTO(client);
     }
 
@@ -37,15 +42,27 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-        Client entity = repository.getReferenceById(id);
-        copyDtoEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Cliente inexistente");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new DatabaseException("Falha Integridade");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha Integridade");
+        }
     }
 
     private void copyDtoEntity(ClientDTO dto, Client entity) {
